requirements.txt
python-dotenv==1.1.1
langchain==0.3.27
langchain-google-genai==2.1.10
google-generativeai==0.8.5
langchain_community==0.3.29
faiss-cpu==1.12.0
langchain-text-splitters==0.3.11
pymupdf==1.26.4